Filetransfer port is u16

+++ src/filetransfer/mod.rs
@@ -72,7 +72,7 @@ pub trait FileTransfer {
     /// 
     /// Connect to the remote server
 
-    fn connect(&mut self, address: String, port: usize, username: Option<String>, password: Option<String>) -> Result<(), FileTransferError>;
+    fn connect(&mut self, address: String, port: u16, username: Option<String>, password: Option<String>) -> Result<(), FileTransferError>;
 
     /// ### disconnect
     ///
